The compilation errors
ShaderManager.h - Changed uint32_t to DWORD in RenderBatch struct (lines 40-41)
ShaderManager.h - Changed unsigned int to DWORD in ExecuteBatches signature
ShaderManager.cpp - Updated ExecuteBatches implementation to use DWORD
SpriteRenderer.cpp - Changed cast to DWORD for primitiveCount
SpriteRenderer.h - Changed uint32_t to DWORD in DrawAtlasSpritePublic
SpriteRenderer.h - Moved FillStagingAreaPublic and getter methods to public section

+++ Settlers2/Graphics/SpriteRenderer.cpp
@@ -767,7 +767,7 @@ void SpriteRenderer::Flush(ShaderManager* pShader) {
         batch.pShader = pShader->GetShader(m_currentShaderName.c_str());
         batch.shaderName = m_currentShaderName;
         batch.startVertex = 0; // Always 0 for now (single buffer)
-        batch.primitiveCount = m_spriteCount * 2; // 2 triangles per sprite
+        batch.primitiveCount = (DWORD)(m_spriteCount * 2); // 2 triangles per sprite
         
         // Set default render states for 2D sprites
         batch.states.zEnable = D3DZB_FALSE;
